Explicit declaration of input box location
html.find("form button") is too generic. If any other module adds a button to the Wall Config dialog, this find would return all buttons and then add multiple wall height boxes.

The proposed changes here allow for a specific location to be declared for the form groups to be added. As the original placed the wall heights at the end of the dialog, this convention has been kept here.